Handle importing images / docs / child collections when no mapping exists for the root collection
We cannot guarantee that an IDMapping exists at the destination site for the root collection, and if it doesn't, we don't want the import to create a new root node (or, worse, fail to create one by trying to call add_child on None). Instead, if we encounter a collection (or any tree-based model) with a null parent_id, map it to the destination site's existing root node.