Eliminate file transfer as a distinct objective/task/operation type
Because file transfer operations have no dependencies of their own, they can be easily inlined into the point where we process the FileField in the corresponding model. We keep ImportContext.imported_files_by_source_url as a cache, so that multiple FileFields pointing to the same file are still handled efficiently.

Doing this means that we're back to objectives always dealing with the life-cycle of database objects, which paves the way for us to do more useful things with them in the forthcoming refactor (such as indexing them in a dict based on object ID).