Implement field-based object lookups
Up to now, "UID" has been synonymous with the UUID string seen in the 'mappings' section of an API response, and stored in the IDMapping table, to identify already-existing objects on the destination site. We now broaden that definition to mean any sort of identifying value that a Locator object can use for that purpose.

We define a new locator type, FieldLocator, which can be enabled for a given model by specifying a list of fields in the WAGTAILTRANSFER_LOOKUP_FIELDS setting. When this is in use, the UID is a tuple of field values (or a list in the JSON representation) and lookups are performed by querying the model itself rather than the external IDMapping table.