Ensure imported pages are added to revision history
If we don't do this, and the page has draft edits at the destination, then the 'edit page' view will continue to treat those draft edits as more recent than the imported (published) version of the page. Additionally, saving a revision has the side effect of updating the Page.draft_title field with the new title, so we don't end up with an outdated title being shown in listings in the admin.