Handle FileFields with non-callable upload_to properties
The current FileField handler attempts to generate a local filename by calling upload_to as a function, which works for Wagtail's own Image and Document models (which set upload_to to a callable) but not the general case where it can be a string or None. FileField provides an internal generate_filename method which handles all cases - use this instead.