Fix problem transferring a StreamField block featuring a model that uses MTI
This changeset fixes a KeyError when a StreamField features a block which itself involves a model that uses multi-table inheritance

While the Wagtail Transfer docs point out that the `preseed_transfer_table` management command and the `WAGTAILTRANSFER_UPDATE_RELATED_MODELS` should not reference models that use multi-table inheritance, it is still possible to use such models with W-T, as long as the concrete superclass is the one that is mentioned. However, the ChooserBlockHandler did not support the notion of MTI, and instead assumed the model was a single-table concrete model.

The change to use `get_base_model` brings the code into line with the approach elsewhere in the codebase.

(This solution is actually from @jacobtm - I'm just implementing it because I need it sooner than he has time for -- thanks Jacob! :o) )