Avoid migration dependency on Wagtail 2.12
The 0003_permissions migration has a dependency on wagtailcore migration 0060, which was added in Wagtail 2.12, so this fails against older versions. It's not clear why this dependency is needed at all, but out of caution I'm changing it to 0040 (present in all 2.x releases) to ensure that at least the initial wagtailcore migrations have completed.