Fix issue with query string parsing
This issue was leading to incorrect message digests which led to a 403
forbidden error.

The query parameter 'fields' needed to be decoded before use. Otherwise,
it would be encoded twice somewhere inside the request.META.get
(i think -unconfirmed)

i.e
,(comma) -> %2C -> %252C

This extra encoding caused the message digests to mismatch leading to
the 403 error.