add logging for ManyToOneRelAdapter
Knowing which adapters are instantiated and that related objects
are found or not can be useful. The details of the logging style and
where could probably be improved

The idea here is to include useful logging info but perhaps not
too much noise. These logs can help trace if a relation isn't followed
as expected.